Version 1
This is code to be added to the 'sub-actions of an action in streamerbot. Right click sub-action > core > C# > execute C# code.

This code takes user input (Rawinput) from the twitch reward redeemed and then checks for basic color keys in the dictionary. If they match, then the value will be assigned to rawInput to be used for color tinting.

Additionally, hex values can be entered manually, though if neither hex value nor color (from the dictionary) is found, points will be refunded.